fix : fix button component
